make client errors printable
